requirements.txt
flask
pillow
torch
torchvision
flask_sqlalchemy
flask_login
werkzeug 